use .match() method
Strings have a .match() method, which takes a regex argument. .match() will return an array of match results – containing either the first match, or all matches if the global flag is used.

Return the result of calling the .match() method on str and passing your regex variable as the argument. You'll use this match result later on.